use specific int types to avoid cgo errors when compiling on tip
Conflicts:
	callbacks.c
	callbacks.go

+++ enums.go
@@ -90,7 +90,7 @@ const (
 
 // Logging levels.
 const (
-	LogError uint = 1 << iota
+	LogError uint16 = 1 << iota
 	LogWarning
 	LogDebug
 	LogFunction
